refactor: cleaned up navbars
